Serialising of dictionaries done with Data.Binary
This means there are Binary instances for dictionaries and tries now.

We use magic numbers to validate the binary file; they are in
Puzzler.Conf.  Puzzler.Error will store all our exceptions, whenever there
is no choice but to throw an exception.

FindAnagrams updated to read the binary format.

puzzler.cabal updated because we need base >= 4 to get exceptions.